+++ server/lib/server.js
@@ -176,9 +176,14 @@ Server.prototype._isHTTPServerOpened = function() {
 };
 
 Server.prototype._logBesedaStarted = function() {
-    var serverAddress = this.httpServer.address();
+	var logger = this.options.debug ? util : console;
 
-    (this.options.debug ? util : console).log('Beseda started on ' +
-                                              serverAddress.address +
-                                              ':' + serverAddress.port);
+	try {
+    	var serverAddress = this.httpServer.address();
+		logger.log('Beseda started on ' + serverAddress.address +
+	               ':' + serverAddress.port);
+	} catch (e) {
+		logger.log('Beseda started.');
+		logger.log('ERROR: can\'t receive server address!');
+	}
 };